Add managed service domain flag

+++ cypress/integration/plugins/alerting-dashboards-plugin/destination_spec.js
@@ -55,34 +55,35 @@ describe('Destinations', () => {
       cy.deleteAllDestinations();
     });
 
+    if(cy.env("MANAGED_SERVICE_ENDPOINT")) {
+      it('with SNS', () => {
 
-    it('with SNS', () => {
+        // Confirm we loaded empty destination list
+        cy.contains('There are no existing destinations');
 
-      // Confirm we loaded empty destination list
-      cy.contains('There are no existing destinations');
+        // Route us to create destination page
+        cy.contains('Add destination').click({force: true});
 
-      // Route us to create destination page
-      cy.contains('Add destination').click({ force: true });
+        // Wait for input to load and then type in the destination name
+        cy.get('input[name="name"]').type(SAMPLE_DESTINATION_SNS, {force: true});
 
-      // Wait for input to load and then type in the destination name
-      cy.get('input[name="name"]').type(SAMPLE_DESTINATION_SNS, { force: true });
+        // Select the type of destination
+        cy.get('#type').select('sns', {force: true});
 
-      // Select the type of destination
-      cy.get('#type').select('sns', { force: true });
+        // Wait for input to load and then type in SNS topic ARN and role ARN
+        cy.get('input[name="sns.topic_arn"]').type(SAMPLE_TOPIC_ARN, {force: true});
+        cy.get('input[name="sns.role_arn"]').type(SAMPLE_ROLE_ARN, {force: true});
 
-      // Wait for input to load and then type in SNS topic ARN and role ARN
-      cy.get('input[name="sns.topic_arn"]').type(SAMPLE_TOPIC_ARN, { force: true });
-      cy.get('input[name="sns.role_arn"]').type(SAMPLE_ROLE_ARN, { force: true });
+        // Click the create button
+        cy.get('button').contains('Create').click({force: true});
 
-      // Click the create button
-      cy.get('button').contains('Create').click({ force: true });
+        // Confirm we can see the created destination in the list
+        cy.contains(SAMPLE_DESTINATION_SNS);
 
-      // Confirm we can see the created destination in the list
-      cy.contains(SAMPLE_DESTINATION_SNS);
-
-      // Delete all destinations
-      cy.deleteAllDestinations();
-    });
+        // Delete all destinations
+        cy.deleteAllDestinations();
+      });
+    }
   });
 
   describe('can be updated', () => {
